nav success but need move xml from main to fragment
