+++ app/src/main/java/com/professionalandroid/apps/coin_in_us/MainActivity.java
@@ -20,7 +20,7 @@ public class MainActivity extends AppCompatActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                 R.id.navigation_interested, R.id.navigation_coin, R.id.navigation_stock)
+                 R.id.navigation_coin, R.id.navigation_home, R.id.navigation_stock)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
