+++ app/src/main/java/com/professionalandroid/apps/coin_in_us/ui/coinsearch/CoinSearchViewModel.java
@@ -1,16 +1,16 @@
-package com.professionalandroid.apps.coin_in_us.ui.notifications;
+package com.professionalandroid.apps.coin_in_us.ui.coinsearch;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class CoinSearchViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public CoinSearchViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("");
     }
 
     public LiveData<String> getText() {
