+++ app/src/main/res/layout/fragment_interested.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context=".ui.coinsearch.CoinSearchFragment">
     <TextView
         android:id="@+id/text_title"
         android:layout_width="wrap_content"
