+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -3,24 +3,24 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_coin">
-<!--    이걸 바꿨는데...-->
+    app:startDestination="@+id/navigation_home">
+<!-- 이걸 바꿨는데...-->
 
     <fragment
-        android:id="@+id/navigation_coin"
+        android:id="@+id/navigation_home"
         android:name="com.professionalandroid.apps.coin_in_us.ui.home.HomeFragment"
-        android:label="@string/title_coinSearch"
-        tools:layout="@layout/fragment_coin_search" />
-
-    <fragment
-        android:id="@+id/navigation_interested"
-        android:name="com.professionalandroid.apps.coin_in_us.ui.dashboard.DashboardFragment"
         android:label="@string/title_interested"
         tools:layout="@layout/fragment_interested" />
 
     <fragment
+        android:id="@+id/navigation_coin"
+        android:name="com.professionalandroid.apps.coin_in_us.ui.coinsearch.CoinSearchFragment"
+        android:label="@string/title_coinSearch"
+        tools:layout="@layout/fragment_coin_search" />
+
+    <fragment
         android:id="@+id/navigation_stock"
-        android:name="com.professionalandroid.apps.coin_in_us.ui.notifications.NotificationsFragment"
+        android:name="com.professionalandroid.apps.coin_in_us.ui.stocksearch.StockSearchFragment"
         android:label="@string/title_stockSearch"
         tools:layout="@layout/fragment_stock_search" />
 </navigation>
